NOTE fixed a -1 bug with SVG-LP MSE evaluation, and added valid set MSE for rec
Wanted to do this for a long time, we can record the MSE here and see how well validation does over time.

Also I am SURE that we really want a -1. We should be dividing over the number of times we run the computation, which is one LESS than the sum (due to for loops)